Split shared state from component modules
